# NETSCRIPT: Greatly speed up script launching, and remove the limitation unique args per script