--- a/src/Netscript/Environment.ts
+++ b/src/Netscript/Environment.ts
@@ -5,11 +5,10 @@ import { NSFull } from "../NetscriptFunctions";
  * Netscript functions and arguments for that script.
  */
 export class Environment {
-  /** Whether or not the script that uses this Environment should stop running */
+  /** Whether or not the script that uses this Environment is stopped */
   stopFlag = false;
 
   /** The currently running function */
-
   runningFn = "";
 
   /** Environment variables (currently only Netscript functions) */