… doc/source/basicgameplay/scripts.rst
… doc/source/basicgameplay/terminal.rst
… markdown/bitburner.ns.exec.md
… markdown/bitburner.ns.isrunning.md
… markdown/bitburner.ns.kill_1.md
… markdown/bitburner.ns.md
… markdown/bitburner.ns.run.md
… markdown/bitburner.runoptions.md
… markdown/bitburner.runoptions.preventduplicates.md
… src/Server/Server.ts
… src/Server/ServerHelpers.ts
… src/Server/ServerPurchases.ts
… src/Terminal/commands/check.ts
… src/Terminal/commands/kill.ts
… src/Terminal/commands/killall.ts
… src/Terminal/commands/ps.ts
… src/Terminal/commands/runScript.ts
… src/Terminal/commands/tail.ts
… src/Terminal/commands/top.ts
… src/ui/ActiveScripts/ServerAccordions.tsx
… src/ui/ActiveScripts/WorkerScriptAccordion.tsx
… src/ui/React/LogBoxManager.tsx
… src/utils/JSONReviver.ts
… src/utils/helpers/roundToTwo.ts
… src/utils/helpers/scriptKey.ts
… src/utils/v2APIBreak.ts
… test/jest/Netscript/RunScript.test.ts
… test/jest/Save.test.ts
… test/jest/__snapshots__/Save.test.ts.snap