@d0sboots 2023-04-10
> This will possibly break a couple of scripts as they rely on the "do not start again" feature.

Ah, you mean scripts are blindly launching something that might already be running, to guarantee it will run, but they only want one copy?

That makes sense. *sigh* I'll put the new behavior in ExecOptions.