@d0sboots 2023-04-11
Alright, I added a new option and also adjusted docs. I think this is ready now.

The new option `preventDuplicates` is false by default, meaning everyone gets the faster behavior but also it may break some people. (But it's much easier to fix in user code, since it's a small addition to get the old behavior back.) If we want to be conservative, we can start with it true and (maybe) change it later.