@steven-r 2023-04-11
I still like the idea and how to circumvent the new behavior.
But as a user for some years now and running over all struggles with the 2.2 API changes: Can we introduce breaking changes in a non "on the fly" manner?
How about having a release policy to allow breaking changes only on Major releases and keep existing APis "as-is", if not fixing a security incident or a major bug.
I have dozens of scripts that might need an adaption which is something I do only if I have enough spare time.

Maybe it's not the right place to discuss this as part of a PR and we should move this to discord.