@d0sboots 2023-04-11
> How about having a release policy to allow breaking changes only on Major releases and keep existing APis "as-is" , if not fixing a security incident or a major bug.

Personally, I agree with you: I think we're too eager to make breaking changes, on the whole. This coming release *will* be a Major release like you describe, because corporations have been significantly overhauled (along with accompanying API changes), but I don't think that lets us off the hook completely.

The only reason I implemented it as default on (and thus breaking) was because this is a significant performance improvement for everyone, especially people who run a lot of scripts. The opposite would be default-slow, which isn't a situation I want if I can help it. Also, it's confusing: Why does launching scripts sometimes fail, even if you have the RAM? Regular OSes don't work like that.

But, just having the option is a big improvement and would allow savvy players a way to get the speedup, so I'm fine with whatever Snarling decides on this one.