@steven-r 2023-04-11
I agree on the performance aspect and having this change together with the corporation overhaul as a major release.

Then we as the users should have a long enough beta time to adjust our scripts accordingly (or even having to versions of BB active for some time (if steam allows this)