@Snarling 2023-04-13
As far as performance improvements, another way we can accomplish the same thing might be through a clever data structure.

One idea without changing game behavior is that we could make server.runningScripts into a Map keyed by the script's filename. The values of that map would themselves be maps, keyed by the stringified args array with the RunningScript as value.

Launching with an array with no checks is super quick - just push onto the array. With these maps it wouldn't be too much more - it needs to stringify its args array and do a couple of map operations. Still a huge improvement compared to iterating through potentially thousands of runningscripts to filename and then compare all args for matching filenames.

Looking up whether a script is running or not (for instance, to prevent player from deleting a file that is currently running) becomes extremely simple - just check if the runningscripts map has the script filename. Compared to iterating potentially the entire array of runningScripts in the expected case that it's not running and the deletion/moving is allowed.

Finding a particular runningScript with filename+args remains possible, and is now that same stringify + couple of map lookups instead of iterating through a massive array and doing comparisons on every entry.

When a runningScript dies, with an array it would have to splice itself out which is slow on big arrays. With maps, removing itself is super quick, just a couple of map operations. In cases where there are a lot of scripts running, this should average way faster than splicing an array.

Thoughts?