@Snarling 2023-04-13
>  I think "all arguments must be unique" is a bad/confusing limitation.

Yeah to an extent I agree with this. So on a gameplay perspective I could kind of go either way on it - one way to favor stability, the other way to favor a change that makes the game make more sense.

>One thing I'll note is that due to memory layout and access patterns, splicing an element out (if you know it's location) is much faster than searching for it.

Yeah so I did some testing, and splice on its own is surprisingly more performant than I thought it would be, considering all elements after the splice need to have their indices adjusted. Probably some browser optimization magic going on behind the scenes for that. Notable though that a runningScript can't really know its own index because that info is not static, so Search + Splice is the more realistic comparison for how a runningScript has to remove itself from the array. That's true for script death whether duplicates are allowed or not. Allowing duplicates makes scripts quick to launch with arrays, but they will still be slow to die.

Here's the test results for different methods of clearing a large runningScripts array/map one RunningScript at a time:

![image](https://user-images.githubusercontent.com/84951833/231885418-25542750-fe7c-4404-a211-031f51a49978.png)