@Snarling 2023-04-14
> I think I'll just make a single-level map that's directly keyed by JSON.stringify of the script name + args.

I thought about this and I don't think it works for determining if a particular script is running in the first place, which is used right now to prevent deletion or moving of a file that is currently running. With that kind of key, you would need to iterate through all keys and do some string analysis on each one to determine if a particular script is running.

If you want to implement it can you commit to the branch from #473 where we already have JSONMaps? I also don't mind implementing it.